DMPlex: Performance bug fix for facet loop iteration space
As it turns out, the order of the facet-to-dof maps previously ignored
the reordering derived by Firedrake that dictates the data layout of
function spaces. This results in bad cache locality and was caused by
facet maps being generated from DMLabel objects, which sort point
indices internally.

This is now fixed by first deriving a global facet traversal order from
a PetscSection that honours the permutation and then building all
facet-specific maps from the ordering. As a result, the generated
facet sets now follow the derived global reordering and therefore
agree more closely with the generated data layout.