Save MixedFunctionSpace subspace splits
When using MFS.create_subdm, we would previously always recreate a new
DM and IS (along with a MixedFunctionSpace) when asking for the subdm.
Now we remember the split according to which fields were requested.
This should reduce some memory churn when using nested fieldsplits in
nonlinear solvers (where the preconditioner is rebuilt each iteration).