solvers: Add more extensive error checking of arguments
Validate not only that arguments to solve calls have the correct type,
but also the correct value (e.g. the LHS of a linear variational solve
must be a bilinear form).  This results in much easier to understand
error messages if the user does the wrong thing and accidentally
provides (say) a residual on the LHS of a linear variational solve, or
similar.