mesh: Compute merged unique ids of exterior facets
Previously, the labels defining the boundary ids would have the same
stratum indices on all processes, this is no longer the case.  Instead
of relying on this, perform a custom reduction to merge the local ids,
ensuring that we get a consistent set on all processes.

Also simplify logic of exterior_facets property somewhat.