fusion: Avoid useless gathers in HF
Push gathers from the wrapper to the dispatcher function. This way, data
that is only required by the fused loop (and *not* by the base loop), is
only loaded from memory if strictly necessary; that is, if a fused
iteration still has to be executed.

Example: think of a facets loop (the base loop) merged with a cells loop
(the fused loop). We iterate over the facets, and then we check whether
any of the adjacent cells still have to be executed; if and only if that's
the case, the extra data required by these cells is gathered.