Splitting clone/checkout to allow sha1 'branch'
If a sha1 is passed via --package-branch as opposed to a branch name the current behaviour generates a 'remote branch not found' error. To allow a specific sha1 to be passed for building, the initial git clone just clones the repository, and then (via a pushed directory) checks out the branch/sha1 as passed at script invocation. Thanks to David Ham for help identifying and fixing this.