Drone CI support
Adding a .drone.yml file which defines the build recipe and environment
for Firedrake on Drone CI. This uses a Ubuntu Xenial base image, onto
which are installed sudo, python-minimal, and python-pip debs, as well
as a pip install of virtualenv. A firedrake user is added, with sudo
(nopasswd) privileges, and with a /firedrake home directory inside which
firedrake-install is run and testing happens.

In addition, the firedrake-install script is modified to build pull requests
from a git clone as well as a git update, and uses drone/jenkins environment