Allow assembling forms with multiple domains
Plausible route to supporting ALE formulations where we formulate the
problem as:

   L = term1*dx(domain=new_mesh) + term2*dx(domain=old_mesh)

Both meshes must share the same topology for this to work.  All trial
and test functions must be defined on function spaces on only one of the
meshes.  Fails if we write:

   L = TestFunction(old_mesh)*dx(domain=new_mesh)

The issue is that somewhere in UFL the determination of the domains of
the form only looks at the measures.  Then further on some analysis
tries to map the domain of the TestFunction and can't find it.