Correctly implement the "otherwise" subdomain
The "otherwise" subdomain that appears in preprocessed forms is not an
integral over all mesh entities, but rather over all mesh entities that
are not caught by any explicitly enumerated subdomains.  If the form
contains no explicit subdomain ids, then "otherwise" = "everywhere".
If not, "otherwise" = "everywhere" - set(explicit subdomain).  To
implement this, we pass the list of explicitly provided subdomain ids to
measure_set and use this to construct the appropriate "complement"
subset.

Fixes #797 and #281.