install: Automagically frameworkise python
This way you can use both python (and ipython) inside the virtualenv on
MacOS and access the GUI.  This is necessary for the default MacOSX
backend of matplotlib to work inside the virtualenv.

This inlines the solution from
https://github.com/gldnspud/virtualenv-pythonw-osx, but modifies it to
overwrite the versioned python binary.  In this way applications that
are installed via the easy_install "entry script" setup, like IPython,
pip, etc... will all use the framework build.