Rework solver parameter setting
This removes the horrible __del__ methods on LinearSolver and
NonlinearVariationalSolver by, in part, following Matt's dictat that
"SetFromOptions should only ever be called on an object once".

The control flow now is that we take in the parameters the user
provides.  When we call solve on the solver object the *first* time, we
meld these options with those from the PETSc options database to create
the correct set of options.  Then we call XXXSetFromOptions on the PETSc
solver.  When the solve is done, we remove the options from the database
again (so that the options database doesn't overflow), and mark this
object that the options have been set (and therefore shouldn't do so
again).

As part of this, rework the way options_prefix works such that two
solvers with ostensibly the same options prefix but different options
are set correctly, fixing #550.