dmplex: Fix performance regression in marking mesh entities
We create label indices for the different entity classes for fast point
lookup.  However, clearing a value invalidates the index which is then
rebuilt the next time the label is queried.  We followed exactly this
pattern in the marking of non-exec facets.  Which resulted in quadratic
complexity for mesh entity marking in parallel.  Fix this by delaying
clearing values until after we have marked all non-exec facets.