Multigrid inside fieldsplits
Rework how information flows between PETSc and Firedrake callbacks,
sanitising the information we attach to the DMs.  In addition, add the
appropriate splitting of solver contexts inside creation of field
decompositions and DM coarsening such that the resulting DMs know how to
assemble coarse grid operators even inside a fieldsplit PC.  This
facilitates (for example) solving Stokes with a geometric multigrid
preconditioner for the velocity block inside a fieldsplit
preconditioner.

In addition to this, simplify the user interface to building multigrid
solvers.  The user now only has to build a MeshHierarchy.  They then set
up their problem as normal using FunctionSpaces just on the finest mesh
level.  Coarse spaces (and coarse jacobians, etc...) will be created if
necessary inside the PC setup if a multigrid PC is requested.