firedrake-status command.
Create a firedrake-status script which dumps to screen and/or a log
file the current configuration as well as the branch and release of
every component. This is:

a. Useful information for users to provide when reporting problems.
b. Convenient if you can't remember what state your install is in.