Making use of drone 0.5 workspace
Drone 0.5 allows the workspace to be modified, specifying where the
initial git checkout is placed. This commit adds a workspace section to
the yaml, sets GIT_SOURCE_DIR to point to the initial git checkout, and
moves the firedrake install into the workspace in case future yaml
changes need it to be available across multiple containers.