More careful fix of component bcs
If a boundary node appears in one BC object that sets all the
components, but another that only sets one component, we will set some
high bits.  Later, in PyOP2, we use the presence of high bits to decide
which components to kill.  Therefore, if any of the BC objects are going
to set a component, we should "unroll" the full BCs to set all the high
bits so that we kill all the components.  Fixes #920, hopefully for good.