solving: Support nesting of options dictionaries
For deeply nested solvers, the flat options dict is rather hard to
read, and moreover, one can't reuse options (e.g. for an AMG solve) in
different sub-blocks.  To fix this, support specification of options
using nested dicts (conceptually these are -options_prefix_push
-options_prefix_pop pairs).  For example

    {"pc_type": "fieldsplit",
     "fieldsplit_0_": {"pc_type": "hypre",
                       "ksp_type": "richardson"},
     "fieldsplit_1_": {"pc_type": "none",
                       "ksp_type": cg}}

is equivalent to

    {"pc_type": "fieldsplit",
     "fieldsplit_0_pc_type": "hypre",
     "fieldsplit_0_ksp_type": "richardson",
     "fieldsplit_1_pc_type": "none",
     "fieldsplit_1_ksp_type": cg}

but easier to read.