Add error checking inside derivative for u.split()
If one defines a form with u.split(), rather than split(u), and then
attempts to take a derivative wrt u, bad things happen.  In
particular, UFL does not notice that the form has dependence on u, and
therefore the resulting form is always zero.  Since this is an easy
mistake to make, we should catch it.  I extract the coefficients from
the form and check if any of them are in the set of u.split().  If so,
raise a ValueError suggesting a fix.  No false positives occur in the
test suite, so I'm going with a exception, rather than a warning,
because no-one reads those anyway.