Automagically append _ to options prefixes
If the options prefix does not end in "_", automatically add it, so
that we can write:

   {"a": {"b": "c"}}

or

   {"a_": {"b": "c"}}

and have them both expand to

   {"a_b": "c"}

For consistency, also do this with the solver options_prefix, so that

   solve(..., options_prefix="foo")

and

   solve(..., options_prefix="foo_")

both obtain an options prefix of "foo_" so that

   python foo.py -foo_ksp_type cg

works in both cases.

Also add a bit more documentation of this treatment.