Fix parallel determination of layer_extents
Computation of layer extents for halo entities might require cell data
that we don't have.  We are, however, guaranteed to have the correct
data for all entities that we own (since we have an overlapped cell
halo).  In parallel, therefore, to compute the correct halo entity
values, we broadcast over the DM's point SF to gather data from
owners.

This will stop working once we no longer have an overlapped cell halo,
in which case we will need to be smarter by adding a parallel
reduction phase instead.