Move to owner computes model
This necessistates off process entries in matrix assembly being
communicated, and off process contributions in vector assembly
likewise.  As well as this we must rework the sparsity construction.

Now that we have owner computes, and are going to make halos smaller,
we need to do communication during the construction of the sparsity.
Rather than rolling our own, just use the special MatPreallocator
matrix type from PETSc to do this.

This means that the sparsity no longer contains a rowptr/colidx pair,
but most of the time it didn't anyway.