jenkins: Work around shebang line length limits
execve(2) says:

    A maximum line length of 127 characters is allowed for the first
    line in an interpreter scripts.

Jenkins runs the test suite in a very deep directory hierarchy.
The installation of pip and pytest puts binaries in PATH with a
shebang line that is therefore longer than this limit.  As a result,
the interpreter fails to execute them.  To fix this, just run as
"python -m xxx ...", rather than "xxx ...".