Add a couple of tests showing the problems with assembling forms with terms on different meshes.
test_domians.py is the MFE that Lawrence posted in the Slack channel a while ago.

test_movingmeshes.py is closer to what we're trying to do with moving meshes in Gusto. It gives a different error message to test_domains.py but could well be essentially the same error.