zenodo: Overhaul release creation
Now that Zenodo doesn't create new DOIs for github releases with new
tags pointing at existing tagged/released commits, we need a new way
of collating the DOIs for a release.

To fix this, we do two things:

1. Continue tagging and releasing on github as usual.
2. Run a second zenodo release step that creates a collated record
   that provides a single citeable DOI that references all the
   components.

This latter step is done by traversing the github repos finding the
candidate tags that might have triggered a Zenodo release, and
resolving one of them (hopefully) to a DOI.

Having gathered DOIs for all the components, we then create a single
Zenodo record that links to all of those components and upload it.

It is this record that the user then needs to cite.

As an added bonus, because we have control over this record creation,
the user can also provide a free-form text file that will be uploaded
with the release (perhaps containing DOIs of their archived simulation
code).