dmhooks: Rework and cleanup subproblem setup
To both cleanly remove subproblems from nested solvers, and be able to
set them up again for multiple uses, we must save a trail of
breadcrumbs down the setup stack when PETSc sets up the solvers for
the first time. We do this with a hooks object on the solver we're
setting up: the breadcrumb trail is specific to a solver, not a DM.

This allows us to clean up all Firedrake data from subproblem DMs in
such a way that we no longer leak memory in deeply nested/complication
solvers.

It also makes the subproblem application context management less
fragile. The DM can be shared across multiple different solvers, and
the solver is responsible for remembering all the bits that cause the
setup to occur.