slate: Bug-for-bug compatibility with existing compiler
The "otherwise" subdomain is treated incorrectly (as "everywhere"),
but fixing it now would be too much work: we need to have a
cell_facets array that is populated with "-1" on "otherwise" faces for
every set of integer subdomain ids.

Look at how it is handled in Facet.measure_set (mesh.py) for
inspiration.