zenodo: don't create multiple releases on same commit
GitHub have removed the ability to create a release for a given commit
if there is already a release referencing that commit. So check for
that case, and if it happens, just create a normal tag instead.

This works fine because the resolution of tags to commits and hence
zenodo releases in create_metarecord walks through tags, not releases.