mg: Move transfer operators when coarsening function spaces
We already move transfer operators when coarsening a DM, but not all
coarse spaces are created by coarsening a DM (e.g. auxiliary
coefficients on a different space in MG). Hence we must replicate some
code: bad design, but the callbacks would be mutually recursive if we
tried not to (at least without major refactoring).