dmhooks: Refactor DM management on functionspaces
It was always a hack to share a DM between all functionspaces with the
same data layout. Change things so that we only share a DM when it's
really the same function space.

This results in separate DMs for topological and geometric
spaces (good!), and also indexed spaces.

This enables us to add hooks to retransfer coefficients to coarse grid
spaces.