dmhooks: Always push the setup_hooks object onto the stack
If the first solve terminates early (for example the residual is zero
due to an initial guess), then the setup of inner objects will not be
run. However, the breadcrumb trail assumed that inner PCs were setup
in the first solve.

To fix this, always push the solver's setup_hooks object onto the
DM (even if we've already run the top-level setup).

Fixes thetisproject/thetis#167.