update LinearVariationalProblem docstring
Clarifies that the `u` parameter is where the solution is assigned when the problem is solved, as opposed to being the function within the input forms for which a solution is sought. (The solution is sought for whatever the `TrialFunction` is in the bilinear form)

This change has been made after getting confused whilst following the "DG advection equation with upwinding" example.