First implementation of p-multigrid.
It converges nicely for the Poisson equation with CG(8), using
power-of-2 coarsening.

The current major limitation is that it uses the same transfer
routines as firedrake.P1PC, which only work for CG & DG on
simplices (I think).

Another major point to be improved is the cleanup: at the moment
no teardown is performed. (I wanted to get it working first.)