tests: Remove use of badly specified pointwise expressions
In various places we were trying to assign shaped constants pointwise
to intrinsically vector-valued spaces, which was always broken but
happened to work in these cases (because it was just zero).

In a similar vein, remove splitting and broadcasting Constants
shape-wise across mixed spaces (doesn't do the right thing for
intrinsically vector-/tensor-valued spaces).