Fix issue #1841
The function returned by _forward_solve() containing the solution on
replay is used by pyadjoint as the new checkpoint, so we can't just
return the same problem.u every time. Instead copy the solution in
the provided func which is created new by pyadjoint at each solve to
store the solution.

Also, for the initial guess just let the value of problem.u be
overwritten by the checkpoint - this happens automatically as it is also
registered as a dependency - rather than trying to keep the problem.u
value of the previous replay solve. In the typical case where the
initial guess just comes from the previous solve in the forward model,
this should do the same in a full replay. However if, for some reason
the initial guess is set differently in the forward model, the replay
should do the same. This is also more correct in the case of partial
replays (from checkpointing).