Change to MixedVectorSpaceBasis with components in multiple subspaces.
Previously MixedVectorSpaceBasis([basisA, basisB]) would take the
Cartesian product of the two sets of basis vectors and form a basis for
a nullspace of the full mixed space that way. However the Cartesian
product formed that way does not necessarily form a basis, e.g.:

basisA = [a1, a2, a3] and basisB = [b1, b2]

then

(a1, b1) - (a2, b1) + (a2, b2) - (a1, b2) = (0, 0)

so basisA x basisB is not independent.

Instead what is needed - at least what I think a reasonable
interpretation of what MixedVectorSpaceBasis([basisA, basisB]) means - is
the direct sum of span(basisA) and span(basisB), understood in the sense
of: span(basisA) x {0}_B + {0}_A x span(basisB) for which the basis is
simply: basisA x {0}_B + {0}_A x basisB - i.e. the n/o vectors in the
combined basis is just the sum of the number in the individual ones.

Note that this implicitly assumes that a1 is a null vector of the
first diagonal block *and* (a1, 0) is a null vector of the full mixed
system. Although this is true for many applications, it is not for many
others.