cython: fix construction of coarse_to_fine_nodes on extruded hierarchies
When we implemented extruded hierarchies that refined in the extruded
direction, we correctly fixed the restriction maps to notice those
extra cells, but did not do so for injection. As a result, the
extruded direction was always treated as having the same number of
layers as on the coarse grid. Adapt that by looping over the refined
cells in the extruded direction too.