Allow variable-height mesh extrusion
The "layer_height" parameter for mesh extrusion was a single scalar
that gave a uniform thickness for all extruded layers. This is
extended to allow for an array to be used instead, which gives the
thickness for the individual layers.

Unit tests have been added to ensure that the resulting mesh volume is
consistent with the supplied layer heights, and this is also tested in
the case of variable layer extrusion.

We also make sure to document and enforce variable layer heights in
make_extruded_coords, and ExtrudedMesh. The value should either be a
scalar or a 1D array; in case of a higher-dimension array we throw a
RuntimeError in make_extruded_coords. If variable layer heights are
provided, there also need to be enough to cover the maximum number of
extruded layers over the entire mesh.