Sharing DM means sharing residual function.
The `DM` is shared between all solvers and so is its `DMSNES` object. This means that the residual function `DMSNESSetFunction` is also shared between all solvers. For the case of several `NonlinearVariationalSolver` it is not a problem because the only difference between solvers is encapsulated in the context `appctx`. It is a problem when using a `TS` that also shares the same `DM`, as it will steal the `DMSNESSetFunction` and set up its own. When an independent and previously setup `NonlinearVariationalSolver` wants to solve again, it finds that its residual function is no longer there, but `TS`'s is.