Ensure ref coords dat.data is always accessed
Make sure that, even if there are no points, we access dat.data which is
a collective operation (i.e. one which should be performed on all
ranks).
In this case, accessing dat.data changes dat.halo_valid from True to
False which is important for determining correct parallel execution
(e.g. where, on one rank, we have no points and on another multiple
points).