Allow for interpolation kernels dropping arguments
When an interpolating an expression which becomes ufl.constantvalue.Zero
after TSFC calls
expression = ufl_utils.preprocess_expression(expression, complex_mode=complex_mode)
in compile_expression_dual_evaluation the runtime tabulation variable
argument (supplied via the rebuild function in interpolation.py) is not
included as a kernel argument by TSFC's fem.compile_ufl.

To allow for this we now explicitly look for this variable name in the
kernel arguments before adding the necessary coefficient to the
coefficient list that will (eventually) make it into the to the parloop
arguments list.

As noted in the comments, this is a stopgap until
compile_expression_dual_evaluation and compile_python_kernel provide a
better API for mapping coefficients and allow a coefficient to be
supplied alongside the to_element (i.e. the dual space we interpolate
onto).