Allow real interpolation onto V.O.M.
Note that interpolator creation is still not working. Simply
passing the cell_node_map as None to the PyOP2 sparsity (copying what
happens when you interpolate) doesn't work. We gets a failure to
create a PyOP2 kernel after calling compile_expression_dual_evaluation
with a loopy type error:

TypeError: unsupported type for persistent hash keying: <class 'complex'>

This is the same error as is seen elsewhere, e.g. for RTCE and RTCF