Fix adjoint with consecutive nonlinear solves
Similar to fix in dF/dm term, also in dF/du the u that we are solving
for should be replaced by the checkpoint of the _solution_ (i.e. the
output) in the forward model not by the checkpoint of the initial guess.
This only becomes relevant when F is actually nonlinear in u.

Bug introduced in #2201 , which has broken Thetis adjoint

Added regression test that tests a sequence of actual nonlinear solves
which fails without the fix.