codegen: Fix issue with permuted maps
If we have a parallel loop with

   map1 = Map(...)
   map2 = PermutedMap(map1, ...)

   par_loop(..., d1(..., map1), d2(..., map2))

The codegen should only require one global argument (for the data from
map1) since map2 is a local permutation of the global argument. This
was being provided by runtime argument passing, but not by the code
generator, which treated map1 and map2 as distinct things.

To fix this, create a new PMap wrapper for Map objects in the codegen
builder that just know how to index themselves through their
permutation. Now we can share the underlying global map data between
the two local map accesses.