install: Fix determination of BLAS location on MacOS
Previously we shoved the location in config["with_blas"] if we were
asking firedrake-install to install openblas. That didn't update the
blas_location variable that is inspected later to set up the BLAS
environment config, so we would never actually use the information we
determined. Log this information so that we can check things.

In addition, since the MacOS install unconditionally requires
homebrew (even with --no-package-manager), automatically determine the
BLAS location even when --no-package-manager is passed. Error early if
this does not resolve to an actual installed BLAS.