Vertex only mesh tolerance (#2449)
Roundoff error can sometimes cause spurious loss of vertices when constructing a VertexOnlyMesh. This PR enables the user to work around this by setting the tolerance by which the local coordinates of a point will be allowed to lie outside a cell while still counting that point as within that cell.