Take tolerance out of locate_cell signature
It's defined in pointquery_utils.py so by adding it as an argument to
locate_cell it's being defined twice. Avoiding defining it twice would
mean changing a lot of code so I've decided to leave it defined where it
is. I have, however, made it a RealType rather than assuming its double.