Pass tolerance to locate_cell for out_of_mesh_point
If you have more ranks than points then you have to specify an out of
mesh point to locate_cell to make sure you run the same kernels on all
ranks. However, when we don't use the default tolerance, this was not
being passed to the out of mesh point causing the kernels to differ
between ranks. This was already an issue on master but my new tests
showed it for the first time.