Make cell_parent_cell_list consistent for extruded meshes
The cell numbering now matches that for a non-extruded mesh. This,
however, is seemingly not enough information to get the correct
result from interpolation. The trouble is that locate_cell has been
seet up to give these consistent results for extruded and non-extruded
mesh numbering, but I need to work out how interpolate works in the
extruded case.