Unify tolerance as param of MeshTopology/Geometry
The tolerance is now a parameter of MeshTopology and MeshGeometry, where
changing the value for MeshGeometry will cause the spatial index to be
rebuilt if necessary. We use this in Function.at and VertexOnlyMesh
and state that changing the tolerance from its default value (i.e. that
of the mesh) will overwrite the tolerance of the mesh and cause the
spatial index to be rebuilt.