Implement VOM voting algorithm
This introduces a global index numbering and rank labelling for DMSwarm
points. Note that we can't use the in-build DMSwarm numbering in the
field named 'DMSwarmField_pid' because there's a bug in petsc4py which
makes it inaccessible (it uses PETSC_INT64 which is not in petsc4py).

At the moment we leave halos empty to match existing behaviour and fix
the case where we were getting point duplication.

This also adds point redistribution which fixes #2178.

See code for how the algorithm works

Also includes tests