Fix using relative tolerance for absolute tolerance
This was both wrong, and created a massive performance bug for point
evaluations with large point sets. The list of candidate cells would,
in some cases, extend across most of the cells in the mesh! With this
fix I can get a ~65x speed up to `locate_cell_and_reference_coordinate`
in some cases.