Add input_ordering property that makes a VOM with original decomposition
This requires us to allow points in halos for _pic_swarm_in_mesh.

We also make VertexOnlyMeshes nameable

We save the rank that a point is on as a new "onrank" field

We have to keep points which are outside of the mesh when we initally
embed them, but we don't use these points when we create the inital VOM,
just the input_ordering VOM.

Note that we don't have an input_ordering of an input_ordering VOM.

get rid of special code for making a VOM in a VOM

variables have been renamed for clarity.