Less special-case code for interpolation with input ordering
I now create a fake petsc mat which can be used in the top level
interpolation code. I also note that interpolations usually do not
include halo cells, so I make that the case here. As part of this I've
added a new input_ordering_without_halos_sf property alongside the
existing input_ordering_sf property. Note that the input_ordering_sf
property is now not used anywhere in the code, so is a candidate for
removal.