Lower default mesh tolerance to 0.5 from 1.0
I kept having to lower it to get certain meshes to behave as expected
(immersed manifolds for example) and the law of least surprise suggests
the default should be low enough to work for most meshes. 0.5 is still a
reasonable relative tolerance (when compared to cell size) in my
opinion.

My theory about why this happens: Say we make a
UnitIcosahedralSphereMesh() which has 20 cells and radius 1.0. Point
evaluation onto immersed manifolds uses a bounding box expanded to a
3-hypercube of side length aproximately 1.0*cell_diameter, which will
extend well into the sphere. Points inside the sphere will often
therefore be attached to the wrong cell when `locate_cell` is called.

I also remove all the places I lower the tolerance to 0.5 in the tests.