Fix cross-mesh interpolation for radially extruded spheres
I hadn't accounted for tensor product elements not always having the
same sub elements, as ocurrs when, say, interpolating onto a function
space on a radially extruded IcosohedralSphereMesh. In that case the
sub elements of the tensor product element are an interval element and
a triangle element.

I fix this and test for it