Test (and fix!) cross mesh interpolation annotation (#3145)
Allow cross-mesh interpolators with expressions with arguments

I was under the mistaken view that I could not create a cross-mesh
interpolator with an expression that had arguments. The reason for this
mistaken view is that interpolation between vertex-only meshes (VOMs)
and their input-ordering VOMs do not allow this.

Having looked at the maths, I realise that the first step, interpolation
from the source mesh onto the immersed VOM, deals with anything we do to
arguments in an expression. The second step, sending data from the VOM
to the input-ordering VOM, doesn't need to know anything about our
original expression other than the locations of the points. To prove
this I enable the necessary tests.

This is needed to allow pyadjoint to work with mesh-to-mesh
interpoalation.

* Test cross-mesh interpolation annotation

Taylor tests added for both 1st order and hessian calculations

---------

Co-authored-by: David A. Ham <david.ham@imperial.ac.uk>