AssembledPC: Keep original operator (#3330)
AssembledPC is built from the preconditioner bilinear form Jp if supplied in the SNES context. For this cases, the original operator J was removed from the new SNES context. This is problematic if one wants to apply multigrid or a Schur complement fieldsplit on the original operator J using Jp to only build the relaxation / Schur complement approximation, while being able to update residuals correctly with J. This PR attaches the original operator to the new SNES context by default.