Big cleanup of simplex closure ordering code
Ideally this will all get ripped out or moved if #3332 goes anywhere.
For now I have simply tried to minimise the footprint of the niche
reordering code. This has entailed rewriting a fair bit of the algorithm
and it is a bit slower since we have to undo some renumbering and this
requires Python, not Cython.

Tests appear to be passing with segfaulting.