Begin attempting to refactor element logic
Basically I need to be able to specify a particular layout for tensor
product cells (quads) that interleaves the DoFs. This is given by
entity_dofs, but determining the right way to put things together is
challenging. My current approach feels very much like I am
reimplementing a lot of the logic that I already have in pyop3 for
expanding loops.