Absolutely disgraceful hackery to get exterior facets working
The problem is that looping over exterior facets is already looping over
a subset/slice of an axis. If we create a local facet dat for exterior
facets then we have to index it with the "local" index (``index.i``)
because we want to treat *exterior facet* 0 as the same as *facet* 0
(more or less).

This crashes and burns when we want to loop over subsets of the exterior
facets.

I suppose a cleaner solution would be to produce a local facet dat for
each set of subdomains.