Facets are finally working properly
The problem was that the sparsity I was creating was too tight. I have
abandoned my prior approach for sparsity construction and gone for
something much simpler. Now to preallocate a matrix we just run over the
same loops for a prealloc matrix and call set values.

It is frustrating how similar the code for interpolation and assembly
is. Eventually I would like to unify the code paths.