Mixed assembly passing
I now need to fix a bug with high priority: dat.data now returns a
*copy* (it's unavoidable in certain circumstances). This is an issue
because modifying it will not modify the original data. The solution is
to catch the "slice-able" cases and work for those but only allow (with
a warning) .data_ro access and error otherwise for the other cases.