Do not copy nullspaces for MassInvPC
The mass matrix itself does not have a nullspace.

As an example, solving Stokes in a closed box which has a pressure
nullspace. The implicit Schur complement matrix of fieldsplit_1 has that
constant nullspace attached. At the level of fieldsplit_1_Mp however we
have the explicit mass matrix approximation. If we copy the nullspace
from the parent explicit matrix, we get into trouble when we specify a
solve, "fieldsplit_1_Mp_pc_type": "ksp", with that mass matrix
as the nullspace will now be applied within the iterations
of the mass matrix solve, where it does not have one. This leads to hard
to detect convergence failures. If we don't copy the nullspace over (as per
this commit), the nullspace is still removed (as we want)
immediately _after_ the MassInvPc as a whole has been applied, in the
outer loop of the Schur complement solve.