Add a bunch of new intermediate types
I think that NonEmptyAssignment is maybe a bit redundant. I instead need
to change the array types themselves.

Aside: is it a good idea to represent everything with types? I think so
because it means that unexpected state will throw errors. However, it
does raise the question:

Say we have an assignment involving a matrix with some zero-sized
branches, do we represent this as a `ConcretizedMat`, or a
`ConcretizedMat` *plus* a `NonEmptyAssignment`? Should the
`NonEmptyAssignment` be put into a `NonEmptyLoop` and so on?

We have distinct phases during parsing and I don't know if the full type
system should be prepared to accept this.

Also, how do we represent this additional type information. We want to
use interfaces/abstract classes so are they mixins? What inherits from
what?